refactor: Extract 7 more sub-clients (templates, uploads, backups, emails, id-mappings, activity, productivity)
Fourth batch in the TodoistApi domain-decomposition series (follows
#561, #562, #563). Migrates 16 more methods onto the sub-client pattern
across 7 new clients — completing every domain except workspaces, which
stays on TodoistApi for a final standalone PR.

Shared-type prep (minimal, scoped to this refactor):

- `FileResponse` moves from `todoist-api.ts` → `types/http.ts` so the
  two clients that return it (UploadClient, BackupClient) can both
  reference it without circular imports. Still exported from the
  package via `types/index.ts` → `src/index.ts`, so no external break.
- `headersToRecord` helper moves to `src/utils/headers.ts` for reuse
  by UploadClient and BackupClient.

New clients:

- `src/clients/template-client.ts` — 5 methods (export as file/URL,
  create/import from file, import from id). Absorbs the private
  `validateTemplateResponse` helper that stitches together
  project/section/task/comment validators.
- `src/clients/upload-client.ts` — 3 methods (uploadFile, deleteUpload,
  viewAttachment). Preserves the todoist.com domain check on
  viewAttachment and the custom-fetch vs native-fetch branching.
- `src/clients/backup-client.ts` — 2 methods (getBackups,
  downloadBackup). Same fetch-branching pattern as viewAttachment.
- `src/clients/email-client.ts` — 2 methods (email forwarding
  get/create and disable).
- `src/clients/id-mapping-client.ts` — 2 methods (getIdMappings,
  getMovedIds).
- `src/clients/activity-client.ts` — 1 method (getActivityLogs).
  Retains Date→YYYY-MM-DD conversion and
  normalize/denormalize legacy-objectType plumbing.
- `src/clients/productivity-client.ts` — 1 method (getProductivityStats).

`TodoistApi` gains 7 more private fields + ctor instantiations; 16
existing public methods collapse to one-line delegations. Public API,
JSDoc, and test files are unchanged — all 479 tests pass.

`src/todoist-api.ts`: 2366 → 2045 lines.

Remaining on `TodoistApi`: workspace methods (~18), plus `getUser` and
`sync` which will likely stay as the facade's core methods.
